refactor(tools): give the semantic layer its own surface, off lfx-lens
The semantic layer tools now call the dbt Semantic Layer in process
through internal/dbtsl, instead of proxying through lfx-lens over HTTP.

internal/tools/lens.go held both surfaces in 664 lines. It shrinks to
query_lfx_lens alone, which still routes through lfx-lens and keeps its
LFXMCP_LENS_API_* settings. Everything else moves to semanticlayer.go
with its own config type. The tool descriptions and help texts move
verbatim, since they carry prompt engineering that took three
iterations to land and three tests guard them.

The dbt client is constructed in a top-level block in main.go rather
than nested inside the LFX API guard that wraps the lens client: it
authenticates with a static service token and has no Auth0 dependency,
so nesting it there would leave it unconfigured for an unrelated
reason. It is deliberately not given the serviceapi debug transport,
which dumps the Authorization header and would print the dbt service
token into production logs, where debug traffic is currently on.

project_slug is removed from the query tool rather than kept as an
accepted no-op. Its description promised the where clause was
'validated against that foundation's subtree', and with the Snowflake
scope check gone nothing enforces that. A parameter that silently does
nothing while reading as a scoping guarantee is the same class of
failure as a filter value that returns zero rows instead of an error.
Scoping now happens in the where clause like any other filter, which
is how it always actually worked. This freed 69 bytes of the query
tool's 2048-byte description budget.

Issue: LFXV2-2940
