+++ internal/tools/csv.go
@@ -1,0 +1,43 @@
+// Copyright The Linux Foundation and contributors.
+// SPDX-License-Identifier: MIT
+
+package tools
+
+import (
+	"encoding/json"
+	"strings"
+)
+
+// parseCSV splits a comma-separated string into trimmed, non-empty values.
+// Also handles JSON-encoded arrays (e.g. `["a","b"]`) that some MCP clients send.
+func parseCSV(s string) []string {
+	s = strings.TrimSpace(s)
+	if s == "" {
+		return nil
+	}
+	// Handle JSON array strings from clients that serialize arrays as strings.
+	// The ReplaceAll handles double-encoded strings with escaped quotes (e.g. `[\"a\",\"b\"]`).
+	if strings.HasPrefix(s, "[") {
+		cleaned := strings.ReplaceAll(s, `\"`, `"`)
+		var arr []string
+		if err := json.Unmarshal([]byte(cleaned), &arr); err == nil {
+			out := make([]string, 0, len(arr))
+			for _, p := range arr {
+				p = strings.TrimSpace(p)
+				if p != "" {
+					out = append(out, p)
+				}
+			}
+			return out
+		}
+	}
+	parts := strings.Split(s, ",")
+	out := make([]string, 0, len(parts))
+	for _, p := range parts {
+		p = strings.TrimSpace(p)
+		if p != "" {
+			out = append(out, p)
+		}
+	}
+	return out
+}
